refactor: db handlers
